# v2/openapi/fb-unified-openapi.yaml
openapi: 3.0.0
servers:
  - description: Local server
    url: http://0.0.0.0:8000
info:
  title: Fireblocks Provider Connectivity API v2
  version: 0.5.0
  x-logo:
    url: doc-assets/fireblocks-logo-lockup-network-navy.png
    altText: Fireblocks logo
  contact:
    name: Fireblocks
    url: https://fireblocks.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: 
    $ref: README.md
...
paths:
...
  /accounts/{accountId}/capabilities/transfers/withdrawals:
    get:
      operationId: getWithdrawalMethods
      summary: Get list of supported withdrawal methods
      description: |
        Retrieves the list of supported withdrawal methods available for the specified account. Shows which withdrawal types, networks, and destinations are supported for fund transfers.
      tags:
        - capabilities
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of withdrawal methods for account.
          content:
            application/json:
              schema:
                type: object
                required:
                  - capabilities
                properties:
                  capabilities:
                    type: array
                    items:
                      $ref: '#/components/schemas/WithdrawalCapability'
              examples:
                allUsdcToBucket:
                  summary: USDCs withdrawals
                  description: |-
                    This response indicates that all USD and USDC bucket balances could be withdrawn using the USDC tokens on the various blockchains. The USD balance  can be specified directly while token and bucket balances must be defined  in `GET /capabilities/assets` and referenced using their IDs.
                    The asset IDs in this example are the same IDs used in the examples of `GET /capabilities/assets`.
                  value:
                    capabilities:
                      - id: 16b8b2c3-bd61-4745-9c48-3d30c2bc6907
                        withdrawal:
                          asset:
                            assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          nationalCurrencyCode: USD
                      - id: 55f184db-9860-4ed7-9db7-7cded54ef0f4
                        withdrawal:
                          asset:
                            assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          nationalCurrencyCode: USD
                      - id: 68fb13d5-48ff-42d8-8273-1658a9b49082
                        withdrawal:
                          asset:
                            assetId: 4386cf4d-83b2-4410-96da-0d3919a45506
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          nationalCurrencyCode: USD
                      - id: d625f8e7-f8aa-4288-9f39-54fecf024c9a
                        withdrawal:
                          asset:
                            assetId: 360de0ad-9ba1-45d5-8074-22453f193d65
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          assetId: f0844d82-7097-4521-95bc-d843724a893e
                      - id: e1fb5c58-c4c5-4358-9781-1c52edf1f2f7
                        withdrawal:
                          asset:
                            assetId: 606bce6b-ff15-4704-9390-b9e32a6cfcff
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          assetId: f0844d82-7097-4521-95bc-d843724a893e
                      - id: 19e6770f-71a9-483b-9d9f-7054ea2dc832
                        withdrawal:
                          asset:
                            assetId: 4386cf4d-83b2-4410-96da-0d3919a45506
                          transferMethod: PublicBlockchain
                        balanceAsset:
                          assetId: f0844d82-7097-4521-95bc-d843724a893e
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
...
  /accounts/{accountId}/transfers/withdrawals:
    get:
      operationId: getWithdrawals
      summary: Get list of withdrawals sorted by creation time
      description: |
        Retrieves a paginated list of all withdrawal transactions for the specified account. Withdrawals are sorted by creation time and include all types of withdrawal operations.
      tags:
        - transfers
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          $ref: '#/components/responses/WithdrawalsResponse'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/withdrawals/{id}:
    get:
      operationId: getWithdrawalDetails
      summary: Get withdrawal details
      description: |
        Retrieves detailed information about a specific withdrawal transaction, including status, amounts, fees, destination details, and processing information.
      tags:
        - transfers
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/EntityIdPathParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: Withdrawals details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Withdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/withdrawals/blockchain:
    get:
      operationId: getBlockchainWithdrawals
      summary: Get list of withdrawals over public blockchains sorted by creation time
      description: |
        Retrieves a paginated list of withdrawal transactions sent over public blockchains. Includes cryptocurrency transfers to external blockchain addresses, sorted by creation time.
      tags:
        - transfersBlockchain
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of withdrawals.
          content:
            application/json:
              schema:
                type: object
                required:
                  - withdrawals
                properties:
                  withdrawals:
                    type: array
                    items:
                      $ref: '#/components/schemas/BlockchainWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createBlockchainWithdrawal
      summary: Create new withdrawal over public blockchain
      description: Should reject any non blockchain withdrawal request.
      tags:
        - transfersBlockchain
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/X-FBAPI-APPROVED-BY'
        - $ref: '#/components/parameters/X-FBAPI-SIGNED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Withdrawal details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BlockchainWithdrawalRequest'
      responses:
        '201':
          description: New withdrawal has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlockchainWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/withdrawals/fiat:
    get:
      operationId: getFiatWithdrawals
      summary: Get list of fiat withdrawals sorted by creation time
      description: |
        Retrieves a paginated list of fiat currency withdrawal transactions. Includes traditional banking transfers and wire transfers, sorted by creation time.
      tags:
        - transfersFiat
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of withdrawals.
          content:
            application/json:
              schema:
                type: object
                required:
                  - withdrawals
                properties:
                  withdrawals:
                    type: array
                    items:
                      $ref: '#/components/schemas/FiatWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createFiatWithdrawal
      summary: Create new fiat withdrawal
      description: Should reject any non fiat withdrawal request.
      tags:
        - transfersFiat
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/X-FBAPI-APPROVED-BY'
        - $ref: '#/components/parameters/X-FBAPI-SIGNED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Withdrawal details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FiatWithdrawalRequest'
      responses:
        '201':
          description: New withdrawal has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FiatWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/withdrawals/peeraccount:
    get:
      operationId: getPeerAccountWithdrawals
      summary: Get list of withdrawals to peer accounts, sorted by creation time
      description: |
        Retrieves a paginated list of withdrawal transactions sent to peer accounts. Includes transfers to other accounts within the same provider ecosystem, sorted by creation time.
      tags:
        - transfersPeerAccounts
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of withdrawals.
          content:
            application/json:
              schema:
                type: object
                required:
                  - withdrawals
                properties:
                  withdrawals:
                    type: array
                    items:
                      $ref: '#/components/schemas/PeerAccountWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createPeerAccountWithdrawal
      summary: Create new withdrawal to a peer account
      description: Should reject any non peer acount withdrawal request.
      tags:
        - transfersPeerAccounts
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/X-FBAPI-APPROVED-BY'
        - $ref: '#/components/parameters/X-FBAPI-SIGNED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Withdrawal details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PeerAccountWithdrawalRequest'
      responses:
        '201':
          description: New withdrawal has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PeerAccountWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
  /accounts/{accountId}/transfers/withdrawals/subaccount:
    get:
      operationId: getSubAccountWithdrawals
      summary: Get list of withdrawals to sub-accounts, sorted by creation time
      description: |
        Retrieves a paginated list of internal withdrawal transactions between sub-accounts. Includes transfers within the same organization or account structure, sorted by creation time.
      tags:
        - transfersInternal
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/PaginationLimit'
        - $ref: '#/components/parameters/PaginationStartingAfter'
        - $ref: '#/components/parameters/PaginationEndingBefore'
        - $ref: '#/components/parameters/ListOrderQueryParam'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      responses:
        '200':
          description: List of withdrawals.
          content:
            application/json:
              schema:
                type: object
                required:
                  - withdrawals
                properties:
                  withdrawals:
                    type: array
                    items:
                      $ref: '#/components/schemas/InternalWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
    post:
      operationId: createSubAccountWithdrawal
      summary: Create new transfer to another sub-account
      description: Transfers assets between different sub-accounts of the same user. Should reject any withdrawals that are not sub-account transfers.
      tags:
        - transfersInternal
      parameters:
        - $ref: '#/components/parameters/X-FBAPI-KEY'
        - $ref: '#/components/parameters/X-FBAPI-NONCE'
        - $ref: '#/components/parameters/X-FBAPI-SIGNATURE'
        - $ref: '#/components/parameters/X-FBAPI-TIMESTAMP'
        - $ref: '#/components/parameters/X-FBAPI-INITIATED-BY'
        - $ref: '#/components/parameters/X-FBAPI-APPROVED-BY'
        - $ref: '#/components/parameters/X-FBAPI-SIGNED-BY'
        - $ref: '#/components/parameters/SubAccountIdPathParam'
      requestBody:
        description: Withdrawal details
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalWithdrawalRequest'
      responses:
        '201':
          description: New withdrawal has been successfully created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InternalWithdrawal'
        '400':
          $ref: '#/components/responses/BadRequestResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedResponse'
        4XX:
          $ref: '#/components/responses/ErrorResponse'
        5XX:
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
...
    X-FBAPI-KEY:
      name: X-FBAPI-KEY
      in: header
      required: true
      description: API authentication key.
      schema:
        type: string
    X-FBAPI-TIMESTAMP:
      name: X-FBAPI-TIMESTAMP
      in: header
      required: true
      description: Request timestamp in milliseconds since Unix epoch.
      schema:
        type: number
    X-FBAPI-NONCE:
      name: X-FBAPI-NONCE
      in: header
      required: true
      description: Unique identifier of the request.
      schema:
        type: string
    X-FBAPI-SIGNATURE:
      name: X-FBAPI-SIGNATURE
      in: header
      required: true
      description: |-
        Request signature using the chosen cryptographic algorithm. The signature is to be calculated on concatenation of the following request fields in the specified order:
        - `X-FBAPI-TIMESTAMP` - `X-FBAPI-NONCE` - HTTP request method in upper case - Endpoint path, including the query parameters - Request body
      schema:
        type: string
    X-FBAPI-INITIATED-BY:
      name: X-FBAPI-INITIATED-BY
      in: header
      required: false
      description: Conditional. This header is provided only when the corresponding capability requirement is enabled.
      schema:
        type: string
    X-FBAPI-APPROVED-BY:
      name: X-FBAPI-APPROVED-BY
      in: header
      required: false
      description: Conditional. This header is provided only when the corresponding capability requirement is enabled.
      schema:
        type: string
...
    EntityIdPathParam:
      name: id
      in: path
      required: true
      description: Entity unique identifier.
      schema:
        type: string
    SubAccountIdPathParam:
      name: accountId
      in: path
      required: true
      description: Sub-account identifier.
      schema:
        type: string
        pattern: ^[A-Za-z0-9_-]+$
    ListOrderQueryParam:
      name: order
      in: query
      description: Controls the order of the items in the returned list.
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
    PaginationLimit:
      name: limit
      in: query
      description: Maximum number of returned items.
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 10
    PaginationStartingAfter:
      name: startingAfter
      in: query
      description: Object ID. Instructs to return the items immediately following this object and not including it. Cannot be used together with `endingBefore`.
      schema:
        type: string
    PaginationEndingBefore:
      name: endingBefore
      in: query
      description: Object ID. Instructs to return the items immediately preceding this object and not including it. Cannot be used together with `startingAfter`.
      schema:
        type: string
...
  responses:
    BadRequestResponse:
      description: Request could not be processed due to a client error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/BadRequestError'
    UnauthorizedResponse:
      description: Request is unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UnauthorizedError'
    ErrorResponse:
      description: Failed to process request.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/GeneralError'
    WithdrawalsResponse:
      description: List of withdrawals.
      content:
        application/json:
          schema:
            type: object
            required:
              - withdrawals
            properties:
              withdrawals:
                type: array
                items:
                  $ref: '#/components/schemas/Withdrawal'
...
  schemas:
...
    PositiveAmount:
      type: string
      pattern: ^\d+(\.\d+)?
...
    WithdrawalCapability:
      type: object
      description: Capability to withdraw a balance asset using a specific transfer capability. `minWithdrawalAmount` is the minimum amount of the balance currency that can be withdrawn using this capability. If `minWithdrawalAmount` is not specified, it will be assumed that the minimum withdrawal amount is zero.
      required:
        - id
        - withdrawal
        - balanceAsset
      properties:
        id:
          type: string
        withdrawal:
          $ref: '#/components/schemas/TransferCapability'
        balanceAsset:
          $ref: '#/components/schemas/AssetReference'
        minWithdrawalAmount:
          $ref: '#/components/schemas/PositiveAmount'
...
    BlockchainWithdrawal:
      allOf:
        - $ref: '#/components/schemas/WithdrawalCommonProperties'
        - type: object
          required:
            - destination
          properties:
            destination:
              $ref: '#/components/schemas/PublicBlockchainTransaction'
...
    Withdrawal:
      oneOf:
        - $ref: '#/components/schemas/PeerAccountWithdrawal'
        - $ref: '#/components/schemas/InternalWithdrawal'
        - $ref: '#/components/schemas/BlockchainWithdrawal'
        - $ref: '#/components/schemas/FiatWithdrawal'
...
  securitySchemes: {}